feat: use common LogEncoding for rocks_impl and table_kv_impl
