+++ wal/src/rocks_impl/mod.rs
@@ -2,5 +2,4 @@
 
 //! WalManager implementation based on RocksDB
 
-pub mod encoding;
 pub mod manager;
