+++ wal/src/table_kv_impl/encoding.rs
@@ -3,21 +3,11 @@
 //! Constants and utils for encoding.
 
 use chrono::{TimeZone, Utc};
-use common_types::{bytes::BytesMut, time::Timestamp};
-use common_util::{
-    codec::{Decoder, Encoder},
-    config::ReadableDuration,
-};
+use common_types::time::Timestamp;
+use common_util::config::ReadableDuration;
 use table_kv::{KeyBoundary, ScanRequest};
 
-use crate::{
-    kv_encoder::{
-        LogKey, LogKeyEncoder, LogValueDecoder, LogValueEncoder, Result,
-        NEWEST_LOG_VALUE_ENCODING_VERSION,
-    },
-    log_batch::Payload,
-    manager::RegionId,
-};
+use crate::manager::RegionId;
 
 /// Key prefix for namespace in meta table.
 const META_NAMESPACE_PREFIX: &str = "v1/namespace";
@@ -97,67 +87,63 @@ pub fn format_region_key(region_id: RegionId) -> String {
     format!("{}/{}", REGION_META_PREFIX, region_id)
 }
 
-#[derive(Debug, Clone)]
-pub struct LogEncoding {
-    key_enc: LogKeyEncoder,
-    value_enc: LogValueEncoder,
-    // value decoder is created dynamically from the version,
-    value_enc_version: u8,
-}
+// #[derive(Debug, Clone)]
+// pub struct LogEncoding {
+//     key_enc: LogKeyEncoder,
+//     value_enc: LogValueEncoder,
+//     // value decoder is created dynamically from the version,
+//     value_enc_version: u8,
+// }
 
-impl LogEncoding {
-    pub fn newest() -> Self {
-        Self {
-            key_enc: LogKeyEncoder::newest(),
-            value_enc: LogValueEncoder::newest(),
-            value_enc_version: NEWEST_LOG_VALUE_ENCODING_VERSION,
-        }
-    }
+// impl LogEncoding {
+//     pub fn newest() -> Self {
+//         Self {
+//             key_enc: LogKeyEncoder::newest(),
+//             value_enc: LogValueEncoder::newest(),
+//             value_enc_version: NEWEST_LOG_VALUE_ENCODING_VERSION,
+//         }
+//     }
 
-    // Encode [LogKey] into `buf` and caller should knows that the keys are ordered
-    // by ([RegionId], [SequenceNum]) so the caller can use this method to
-    // generate min/max key in specific scope(global or in some region).
-    pub fn encode_key(&self, buf: &mut BytesMut, log_key: &LogKey) -> Result<()> {
-        buf.clear();
-        buf.reserve(self.key_enc.estimate_encoded_size(log_key));
-        self.key_enc.encode(buf, log_key)?;
+//     // Encode [LogKey] into `buf` and caller should knows that the keys are
+// ordered     // by ([RegionId], [SequenceNum]) so the caller can use this
+// method to     // generate min/max key in specific scope(global or in some
+// region).     pub fn encode_key(&self, buf: &mut BytesMut, log_key: &LogKey)
+// -> Result<()> {         buf.clear();
+//         buf.reserve(self.key_enc.estimate_encoded_size(log_key));
+//         self.key_enc.encode(buf, log_key)?;
 
-        Ok(())
-    }
+//         Ok(())
+//     }
 
-    pub fn encode_value(&self, buf: &mut BytesMut, payload: &dyn Payload) -> Result<()> {
-        buf.clear();
-        buf.reserve(self.value_enc.estimate_encoded_size(payload));
-        self.value_enc.encode(buf, payload)
-    }
+//     pub fn encode_value(&self, buf: &mut BytesMut, payload: &dyn Payload) ->
+// Result<()> {         buf.clear();
+//         buf.reserve(self.value_enc.estimate_encoded_size(payload));
+//         self.value_enc.encode(buf, payload)
+//     }
 
-    pub fn is_log_key(&self, mut buf: &[u8]) -> Result<bool> {
-        self.key_enc.is_valid(&mut buf)
-    }
+//     pub fn is_log_key(&self, mut buf: &[u8]) -> Result<bool> {
+//         self.key_enc.is_valid(&mut buf)
+//     }
 
-    pub fn decode_key(&self, mut buf: &[u8]) -> Result<LogKey> {
-        self.key_enc.decode(&mut buf)
-    }
+//     pub fn decode_key(&self, mut buf: &[u8]) -> Result<LogKey> {
+//         self.key_enc.decode(&mut buf)
+//     }
 
-    pub fn decode_value<'a>(&self, buf: &'a [u8]) -> Result<&'a [u8]> {
-        let value_dec = LogValueDecoder {
-            version: self.value_enc_version,
-        };
+//     pub fn decode_value<'a>(&self, buf: &'a [u8]) -> Result<&'a [u8]> {
+//         let value_dec = LogValueDecoder {
+//             version: self.value_enc_version,
+//         };
 
-        value_dec.decode(buf)
-    }
-}
+//         value_dec.decode(buf)
+//     }
+// }
 
 #[cfg(test)]
 mod tests {
     use std::time::Duration;
 
     use super::*;
-    use crate::{
-        log_batch::PayloadDecoder,
-        table_kv_impl::namespace,
-        tests::util::{TestPayload, TestPayloadDecoder},
-    };
+    use crate::table_kv_impl::namespace;
 
     #[test]
     fn test_format_namespace_key() {
@@ -250,35 +236,5 @@ mod tests {
 
         let key = format_region_key(RegionId::MAX);
         assert_eq!("v1/region/18446744073709551615", key);
-    }
-
-    #[test]
-    fn test_log_encoding() {
-        let region_id = 1234;
-
-        let sequences = [1000, 1001, 1002, 1003];
-        let mut buf = BytesMut::new();
-        let encoding = LogEncoding::newest();
-        for seq in sequences {
-            let log_key = (region_id, seq);
-            encoding.encode_key(&mut buf, &log_key).unwrap();
-
-            assert!(encoding.is_log_key(&buf).unwrap());
-
-            let decoded_key = encoding.decode_key(&buf).unwrap();
-            assert_eq!(log_key, decoded_key);
-        }
-
-        let decoder = TestPayloadDecoder;
-        for val in 0..8 {
-            let payload = TestPayload { val };
-
-            encoding.encode_value(&mut buf, &payload).unwrap();
-
-            let mut value = encoding.decode_value(&buf).unwrap();
-            let decoded_value = decoder.decode(&mut value).unwrap();
-
-            assert_eq!(payload, decoded_value);
-        }
     }
 }
